Add new init file
